+++ src/cryptography/hazmat/primitives/asymmetric/ec.py
@@ -5,6 +5,7 @@
 from __future__ import absolute_import, division, print_function
 
 import abc
+import warnings
 
 import six
 
@@ -150,6 +151,31 @@ class EllipticCurvePublicKey(object):
         """
         Verifies the signature of the data.
         """
+
+    @classmethod
+    def from_encoded_point(cls, curve, data, backend):
+        if not isinstance(curve, EllipticCurve):
+            raise TypeError("curve must be an EllipticCurve instance")
+
+        if not backend:
+            raise ValueError('Must provide backend to decode point')
+
+        byte_length = (curve.key_size + 7) // 8
+        compressed_byte_length = 1 + byte_length
+        if (len(data) == compressed_byte_length and
+                data[0:1] in (b'\x02', b'\x03')):
+            data = backend.uncompress_elliptic_curve_bytes(curve, data)
+
+        if data.startswith(b'\x04'):
+            if len(data) == 2 * byte_length + 1:
+                x = utils.int_from_bytes(data[1:byte_length + 1], 'big')
+                y = utils.int_from_bytes(data[byte_length + 1:], 'big')
+                return backend.load_elliptic_curve_public_numbers(
+                    EllipticCurvePublicNumbers(x, y, curve))
+            else:
+                raise ValueError('Invalid elliptic curve point data length')
+        else:
+            raise ValueError('Unsupported elliptic curve point type')
 
 
 EllipticCurvePublicKeyWithSerialization = EllipticCurvePublicKey
@@ -341,18 +367,32 @@ class EllipticCurvePublicNumbers(object):
     def public_key(self, backend):
         return backend.load_elliptic_curve_public_numbers(self)
 
-    def encode_point(self):
+    def encode_point(self, compressed=False):
         # key_size is in bits. Convert to bytes and round up
         byte_length = (self.curve.key_size + 7) // 8
-        return (
-            b'\x04' + utils.int_to_bytes(self.x, byte_length) +
-            utils.int_to_bytes(self.y, byte_length)
-        )
+        if not compressed:
+            return (
+                b'\x04' + utils.int_to_bytes(self.x, byte_length) +
+                utils.int_to_bytes(self.y, byte_length)
+            )
+        else:
+            compression_byte = b'\x03' if self.y % 2 else b'\x02'
+            return (
+                compression_byte + utils.int_to_bytes(self.x, byte_length)
+            )
 
     @classmethod
     def from_encoded_point(cls, curve, data):
         if not isinstance(curve, EllipticCurve):
             raise TypeError("curve must be an EllipticCurve instance")
+
+        warnings.warn(
+            "Support for unsafe construction of public numbers from "
+            "encoded data will be removed in a future version. "
+            "Please use EllipticCurvePublicKey.from_encoded_point",
+            utils.DeprecatedIn24,
+            stacklevel=2,
+        )
 
         if data.startswith(b'\x04'):
             # key_size is in bits. Convert to bytes and round up